split up gmsh_file_v4 so it can be compiled in smaller units
