+++ .travis.yml
@@ -49,7 +49,7 @@ jobs:
     env:
     - COMPILER=g++-7
     - BUILD_TYPE='Debug'
-    - NUM_PROC=1
+    - NUM_PROC=2
     addons:
       apt:
         update: true
