+++ lib/lf/io/CMakeLists.txt
@@ -3,6 +3,9 @@ set(sources
   gmsh_file_v2.h
   gmsh_file_v4.cc
   gmsh_file_v4.h
+  gmsh_file_v4_detail.h
+  gmsh_file_v4_text.cc
+  gmsh_file_v4_binary.cc
   gmsh_reader.cc
   gmsh_reader.h
   io.h
